feat: fix some eslint-ts error, and create/page disable ssr
