+++ src/app/proposal/details/components/VoteInfo.tsx
@@ -1,5 +1,5 @@
 import { Typography, FontWeightEnum, Progress } from 'aelf-design';
-import MyInfo from 'app/daodetails/components/MyInfo';
+import MyInfo from 'app/daodetails/[daoId]/components/MyInfo';
 import BoxWrapper from './BoxWrapper';
 import { memo } from 'react';
