+++ src/constants/index.ts
@@ -1,14 +1,12 @@
 export enum WalletType {
-    unknown = 'unknown',
-    discover = 'discover',
-    portkey = 'portkey',
-  }
-  
-  export enum NetworkType {
-    MAIN = 'MAIN',
-    TESTNET = 'TESTNET',
-  }
+  unknown = 'unknown',
+  discover = 'discover',
+  portkey = 'portkey',
+}
 
-  export const SECONDS_60 = 60000;
+export enum NetworkType {
+  MAIN = 'MAIN',
+  TESTNET = 'TESTNET',
+}
 
-  
+export const SECONDS_60 = 60000;
