+++ src/contract/token.ts
@@ -1,9 +1,9 @@
 import {
-	SupportedELFChainId,
-	IContractOptions,
-	ContractMethodType,
-	ISendResult,
-	IContractError,
+  SupportedELFChainId,
+  IContractOptions,
+  ContractMethodType,
+  ISendResult,
+  IContractError,
 } from 'types';
 import { store } from 'redux/store';
 import { formatErrorMsg } from './util';
@@ -12,59 +12,59 @@ import { sleep } from 'utils/common';
 import { GetContractServiceMethod } from './baseContract';
 
 export const multiTokenContractRequest = async <T, R>(
-	methodName: string,
-	params: T,
-	options?: IContractOptions,
+  methodName: string,
+  params: T,
+  options?: IContractOptions,
 ): Promise<R | ISendResult> => {
-	const info = store.getState().elfInfo.elfInfo;
-	const contractAddressMap = {
-		main: info?.mainChainAddress || '',
-		side: info?.sideChainAddress || '',
-	};
-	const contractAddress = (options?.chain === SupportedELFChainId.MAIN_NET
-		? contractAddressMap.main
-		: contractAddressMap.side) as unknown as string;
+  const info = store.getState().elfInfo.elfInfo;
+  const contractAddressMap = {
+    main: info?.mainChainAddress || '',
+    side: info?.sideChainAddress || '',
+  };
+  const contractAddress = (options?.chain === SupportedELFChainId.MAIN_NET
+    ? contractAddressMap.main
+    : contractAddressMap.side) as unknown as string;
 
-	const curChain: Chain = options?.chain || (info?.curChain as Chain);
+  const curChain: Chain = options?.chain || (info?.curChain as Chain);
 
-	console.log(
-		'=====multiTokenContractRequest methodName, type, contractAddress, curChain, params: ',
-		methodName,
-		options?.type,
-		contractAddress,
-		curChain,
-		params,
-	);
+  console.log(
+    '=====multiTokenContractRequest methodName, type, contractAddress, curChain, params: ',
+    methodName,
+    options?.type,
+    contractAddress,
+    curChain,
+    params,
+  );
 
-	const CallContractMethod = GetContractServiceMethod(curChain, options?.type);
+  const CallContractMethod = GetContractServiceMethod(curChain, options?.type);
 
-	try {
-		const res: R = await CallContractMethod({
-			contractAddress,
-			methodName,
-			args: params,
-		});
-		console.log('=====multiTokenContractRequest res: ', methodName, res);
-		const result = res as IContractError;
+  try {
+    const res: R = await CallContractMethod({
+      contractAddress,
+      methodName,
+      args: params,
+    });
+    console.log('=====multiTokenContractRequest res: ', methodName, res);
+    const result = res as IContractError;
 
-		if (result?.error || result?.code || result?.Error) {
-			throw formatErrorMsg(result);
-		}
+    if (result?.error || result?.code || result?.Error) {
+      throw formatErrorMsg(result);
+    }
 
-		if (options?.type === ContractMethodType.VIEW) {
-			return res;
-		}
+    if (options?.type === ContractMethodType.VIEW) {
+      return res;
+    }
 
-		const { transactionId, TransactionId } = result.result || result;
-		const resTransactionId = TransactionId || transactionId;
-		await sleep(1000);
-		const transaction = await getTxResult(resTransactionId!, curChain as Chain);
+    const { transactionId, TransactionId } = result.result || result;
+    const resTransactionId = TransactionId || transactionId;
+    await sleep(1000);
+    const transaction = await getTxResult(resTransactionId!, curChain as Chain);
 
-		console.log('=====multiTokenContractRequest transaction: ', methodName, transaction);
-		return { TransactionId: transaction.TransactionId, TransactionResult: transaction.txResult };
-	} catch (error) {
-		console.error('=====multiTokenContractRequest error:', methodName, JSON.stringify(error));
-		const resError = error as IContractError;
-		throw formatErrorMsg(resError);
-	}
+    console.log('=====multiTokenContractRequest transaction: ', methodName, transaction);
+    return { TransactionId: transaction.TransactionId, TransactionResult: transaction.txResult };
+  } catch (error) {
+    console.error('=====multiTokenContractRequest error:', methodName, JSON.stringify(error));
+    const resError = error as IContractError;
+    throw formatErrorMsg(resError);
+  }
 };
