+++ src/utils/common.ts
@@ -1,7 +1,7 @@
 export const sleep = (time: number) => {
-	return new Promise<void>((resolve) => {
-		setTimeout(() => {
-			resolve();
-		}, time);
-	});
+  return new Promise<void>((resolve) => {
+    setTimeout(() => {
+      resolve();
+    }, time);
+  });
 };
